Add dom test
