+++ package.json
@@ -14,6 +14,7 @@
     "browserify": "^12.0.1",
     "chrome-webstore-manager": "^0.1.1",
     "crx": "github:pastak/crx",
+    "jsdom": "^7.2.2",
     "request-promise": "^1.0.2",
     "standard": "^4.5.4"
   },
